--- AI4ALL6C/README.md
@@ -1,1 +1,0 @@
-# AI4ALL6C
